Reorg logger and util pkgs, go-lint
